Change location for merge conflict test

See merge request !4

+++ src/cellsociety_team13/Location.java
@@ -3,7 +3,7 @@ package cellsociety_team13;
 import java.awt.Point;
 
 public class Location extends Point{
-	public Location(int x, int y){
+	public Location(double x, double y){
 		
 	}
 }
